move shareable error utils
