+++ webui/react/src/components/ActionDropdown.tsx
@@ -4,9 +4,10 @@ import React, { JSXElementConstructor } from 'react';
 
 import Icon from 'components/Icon';
 import { capitalize } from 'tmp-shared/utils/string';
-import handleError, { ErrorLevel, ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 
 import { Eventually } from '../tmp-shared/types';
+import { ErrorLevel, ErrorType } from '../tmp-shared/utils/error';
 
 import css from './ActionDropdown.module.scss';
