+++ webui/react/src/components/Metadata/MetadataCard.tsx
@@ -3,8 +3,9 @@ import { Button, Card, Space, Tooltip } from 'antd';
 import React, { useCallback, useMemo, useState } from 'react';
 
 import { Metadata } from 'types';
-import handleError, { ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 
+import { ErrorType } from '../../tmp-shared/utils/error';
 import Spinner from '../Spinner';
 
 import EditableMetadata from './EditableMetadata';
