+++ webui/react/src/components/NotesCard.tsx
@@ -3,7 +3,9 @@ import { Button, Card, Space, Tooltip } from 'antd';
 import React, { useCallback, useState } from 'react';
 import { Prompt, useLocation } from 'react-router-dom';
 
-import handleError, { ErrorType } from 'utils/error';
+import handleError from 'utils/error';
+
+import { ErrorType } from '../tmp-shared/utils/error';
 
 import Markdown from './Markdown';
 import css from './NotesCard.module.scss';
