+++ webui/react/src/components/TaskActionDropdown.tsx
@@ -18,9 +18,11 @@ import { capitalize } from 'tmp-shared/utils/string';
 import {
   ExperimentAction as Action, AnyTask, CommandTask, DetailedUser, ExperimentTask, RunState,
 } from 'types';
-import handleError, { ErrorLevel, ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 import { isExperimentTask, isTaskKillable } from 'utils/task';
 import { openCommand } from 'wait';
+
+import { ErrorLevel, ErrorType } from '../tmp-shared/utils/error';
 
 import css from './ActionDropdown.module.scss';
 import Link from './Link';
