+++ webui/react/src/hooks/useTelemetry.ts
@@ -3,7 +3,9 @@ import { useRef } from 'react';
 
 import { getTelemetry } from 'services/api';
 import { Auth, DeterminedInfo } from 'types';
-import handleError, { ErrorLevel, ErrorType } from 'utils/error';
+import handleError from 'utils/error';
+
+import { ErrorLevel, ErrorType } from '../tmp-shared/utils/error';
 
 /*
  * Telemetry is written as a modular class instance instead of
