+++ webui/react/src/ioTypes.ts
@@ -4,7 +4,8 @@ import * as io from 'io-ts';
 import {
   CheckpointStorageType, ExperimentSearcherName, HyperparameterType, LogLevel, RunState,
 } from 'types';
-import { DetError, ErrorLevel, ErrorType } from 'utils/error';
+
+import { DetError, ErrorLevel, ErrorType } from './tmp-shared/utils/error';
 
 /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
 export const decode = <T>(type: io.Mixed, data: any): T => {
