+++ webui/react/src/pages/JobQueue/JobQueue.tsx
@@ -21,10 +21,12 @@ import { ShirtSize } from 'themes';
 import { isEqual } from 'tmp-shared/utils/data';
 import { capitalize } from 'tmp-shared/utils/string';
 import { Job, JobAction, JobType, ResourcePool, RPStats } from 'types';
-import handleError, { ErrorLevel, ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 import { canManageJob, moveJobToPosition, orderedSchedulers,
   unsupportedQPosSchedulers } from 'utils/job';
 import { numericSorter } from 'utils/sort';
+
+import { ErrorLevel, ErrorType } from '../../tmp-shared/utils/error';
 
 import css from './JobQueue.module.scss';
 import settingsConfig, { Settings } from './JobQueue.settings';
