+++ webui/react/src/pages/JobQueue/ManageJob.tsx
@@ -8,8 +8,10 @@ import { updateJobQueue } from 'services/api';
 import * as api from 'services/api-ts-sdk';
 import { floatToPercent, truncate } from 'tmp-shared/utils/string';
 import { Job, JobType, RPStats } from 'types';
-import handleError, { ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 import { moveJobToPositionUpdate, orderedSchedulers, unsupportedQPosSchedulers } from 'utils/job';
+
+import { ErrorType } from '../../tmp-shared/utils/error';
 
 import css from './ManageJob.module.scss';
 const { Option } = Select;
