+++ webui/react/src/pages/ModelRegistry.tsx
@@ -28,10 +28,11 @@ import { V1GetModelsRequestSortBy } from 'services/api-ts-sdk';
 import { isBoolean, isEqual } from 'tmp-shared/utils/data';
 import { capitalize } from 'tmp-shared/utils/string';
 import { ArchiveFilter, ModelItem } from 'types';
-import handleError, { ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 import { alphaNumericSorter } from 'utils/sort';
 import { getDisplayName } from 'utils/user';
 
+import { ErrorType } from '../tmp-shared/utils/error';
 import { validateDetApiEnum } from '../tmp-shared/utils/service';
 
 import css from './ModelRegistry.module.scss';
