+++ webui/react/src/pages/TrialDetails/TrialDetailsLogs.tsx
@@ -13,7 +13,9 @@ import { mapV1LogsResponse } from 'services/decoder';
 import { readStream } from 'services/utils';
 import { ExperimentBase, TrialDetails } from 'types';
 import { downloadTrialLogs } from 'utils/browser';
-import handleError, { ErrorType } from 'utils/error';
+import handleError from 'utils/error';
+
+import { ErrorType } from '../../tmp-shared/utils/error';
 
 import css from './TrialDetailsLogs.module.scss';
