+++ webui/react/src/utils/error.ts
@@ -4,100 +4,13 @@ import { ArgsProps, NotificationApi } from 'antd/lib/notification';
 import { telemetryInstance } from 'hooks/useTelemetry';
 import { paths } from 'routes/utils';
 import history from 'tmp-shared/routes/history';
-import Logger, { LoggerInterface } from 'tmp-shared/utils/Logger';
+import { LoggerInterface } from 'tmp-shared/utils/Logger';
 import { listToStr } from 'tmp-shared/utils/string';
 
-import { isString } from '../tmp-shared/utils/data';
+import { DetError, DetErrorOptions, ERROR_NAMESPACE, ErrorLevel, ErrorType,
+  isDetError } from '../tmp-shared/utils/error';
 import { filterOutLoginLocation } from '../tmp-shared/utils/routes';
 import { isAborted } from '../tmp-shared/utils/service';
-
-export interface DetErrorOptions {
-  id?: string; // slug unique to each place in the codebase that we will use this.
-  isUserTriggered?: boolean; // whether the error was triggered by an active interaction.
-  level?: ErrorLevel;
-  logger?: LoggerInterface;
-  payload?: unknown;
-  publicMessage?: string;
-  publicSubject?: string;
-  silent?: boolean;
-  type?: ErrorType;
-}
-
-export enum ErrorLevel {
-  Fatal = 'fatal',
-  Error = 'error',
-  Warn = 'warning',
-}
-
-export enum ErrorType {
-  Server = 'server', // internal apis and server errors.
-  Auth = 'auth',
-  Unknown = 'unknown',
-  Ui = 'ui',
-  Input = 'input', // the issue is caused by unexpected/invalid user input.
-  ApiBadResponse = 'apiBadResponse', // unexpected response structure.
-  Api = 'api', // third-party api
-}
-
-const ERROR_NAMESPACE = 'EH';
-const DEFAULT_ERROR_MESSAGE = 'Unknown error encountered.';
-const DEFAULT_LOGGER = new Logger(ERROR_NAMESPACE);
-
-export const isError = (error: unknown): error is Error => {
-  return error instanceof Error;
-};
-
-export const isDetError = (error: unknown): error is DetError => {
-  return error instanceof DetError;
-};
-
-const defaultErrOptions: DetErrorOptions = {
-  isUserTriggered: false,
-  level: ErrorLevel.Error,
-  logger: DEFAULT_LOGGER,
-  silent: false,
-  type: ErrorType.Unknown,
-};
-
-// An expected Error with supplemental information on
-// how it should be handled.
-export class DetError extends Error implements DetErrorOptions {
-  id?: string;
-  isUserTriggered: boolean;
-  level: ErrorLevel;
-  logger: LoggerInterface;
-  payload?: unknown;
-  publicMessage?: string;
-  publicSubject?: string;
-  silent: boolean;
-  type: ErrorType;
-  isHandled: boolean;
-
-  constructor(e?: unknown, options: DetErrorOptions = {}) {
-    const defaultMessage = isError(e) ? e.message : (isString(e) ? e : DEFAULT_ERROR_MESSAGE);
-    const message = options.publicSubject || options.publicMessage || defaultMessage;
-    super(message);
-
-    const eOpts: DetErrorOptions = isDetError(e) ? {
-      id: e.id,
-      isUserTriggered: e.isUserTriggered,
-      level: e.level,
-      logger: e.logger,
-      payload: e.payload,
-      publicMessage: e.publicMessage,
-      publicSubject: e.publicSubject,
-      silent: e.silent,
-      type: e.type,
-    } : {};
-
-    this.loadOptions({ ...defaultErrOptions, ...eOpts, ...options });
-    this.isHandled = false;
-  }
-
-  loadOptions(options: DetErrorOptions): void {
-    Object.assign(this, options);
-  }
-}
 
 const errorLevelMap = {
   [ErrorLevel.Error]: 'error',
