+++ webui/react/src/utils/job.ts
@@ -1,8 +1,9 @@
 import { updateJobQueue } from 'services/api';
 import * as Api from 'services/api-ts-sdk';
 import { CommandType, Job, JobType, ResourcePool } from 'types';
-import handleError, { DetError, DetErrorOptions, ErrorType } from 'utils/error';
+import handleError from 'utils/error';
 
+import { DetError, DetErrorOptions, ErrorType } from '../tmp-shared/utils/error';
 import { capitalize } from '../tmp-shared/utils/string';
 
 // This marks scheduler types that do not support fine-grain control of
